requirements.txt
flask
flask-sqlalchemy
python-dotenv
requests
pandas
scikit-learn
numpy
opencv-python
mediapipe
pyautogui
pygetwindow
SpeechRecognition
pyaudio